+++ .github/workflows/pylint.yml
@@ -19,6 +19,9 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install pylint
+        pip install -r requirments.txt
     - name: Analysing the code with pylint
       run: |
         pylint `ls -R|grep .py$|xargs`
+    
+
